feat: audio fallback chain, fractional scores, short+say, version file
- audio: Play tries afplay → ffplay → ffmpeg; PlaySegment tries ffplay → ffmpeg
- rickies: Score fields widened to float64 to support half-point picks (e.g. The EUies)
- intro --short: prints first sentence of actual transcript instead of hardcoded string
- intro/exit --short: now combinable with --say
- versions.txt: version codenames moved to embedded file, no longer hardcoded
- banner: version number and title shown centered under the wordmark
- README: updated audio dependency notes

+++ internal/rickies/rickies.go
@@ -103,7 +103,7 @@ type Pick struct {
 	Text       string `json:"text"`
 	Notes      string `json:"notes,omitempty"`
 	Type       string `json:"type,omitempty"`
-	Score      int    `json:"score"`
+	Score      float64 `json:"score"`
 	Conditions int    `json:"pick-conditions"`
 }
 
@@ -114,7 +114,7 @@ func (p Pick) Hit() bool { return p.Score > 0 }
 // use Correct/Total.
 type HostScore struct {
 	Host    string `json:"host"`
-	Score   int    `json:"score,omitempty"`
+	Score   float64 `json:"score,omitempty"`
 	Correct int    `json:"correct,omitempty"`
 	Total   int    `json:"total,omitempty"`
 }
